Move all routes to index and skip combine routes
