+++ package.json
@@ -38,7 +38,6 @@
     "ioredis": "^2.5.0",
     "koa": "^2.0.0",
     "koa-body": "2",
-    "koa-combine-routers": "^1.0.0",
     "koa-convert": "^1.2.0",
     "koa-pug": "^3.0.0-1",
     "koa-router": "^7.0.0",
